Session and control file correction
Corrected the session handling so the session now updates when user
is moving from route to route, and on page reloads.  Session data
is accurate in the mongo session store, and is returned properly to
the views.

Updated package.json to reflect correct dependencies and author.

Updated .gitignore to remove mongo-test.